requirements.txt
streamlit>=1.38.0
pandas>=2.2.3
plotly>=5.18.0
google-api-python-client>=2.114.0
google-auth>=2.26.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-analytics-data>=0.18.0
searchconsole>=0.0.4
python-dotenv>=1.0.0
numpy>=2.1.0
altair>=5.2.0
requests>=2.31.0
oauth2client>=4.1.3